+++ package.json
@@ -6,8 +6,8 @@
     "@c2fo/components": "^2.0.2",
     "@material-ui/core": "^4.9.13",
     "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
+    "@types/jest": "^25.2.1",
     "babel-plugin-react-intl": "^7.5.5",
     "i18next": "^19.4.4",
     "i18next-browser-languagedetector": "^4.1.1",
@@ -40,5 +40,10 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@testing-library/dom": "^7.5.1",
+    "@testing-library/react": "^9.5.0",
+    "@testing-library/react-hooks": "^3.2.1"
   }
 }
